refactor: split the run history out of the Routines page
automations-page.tsx crossed the size cap once the precheck detail landed on
top of the header work that reached main. The seam is the page's own: the list
half answers "what is scheduled and what can I do to it", the runs half answers
"did it run, and what happened" — a read-only projection with no actions and
no state. formatWhen moves to a third file so the runs block can render a
timestamp without importing the page that renders it.
